ttlinit: move CreateRowLevelTTLScheduledJob function
This allow the function to be used by the declarative schemachanger in a
future commit.

Release note: None
