+++ pkg/sql/BUILD.bazel
@@ -546,7 +546,7 @@ go_library(
         "//pkg/sql/syntheticprivilege",
         "//pkg/sql/syntheticprivilegecache",
         "//pkg/sql/tablemetadatacache/util",
-        "//pkg/sql/ttl/ttlbase",
+        "//pkg/sql/ttl/ttlinit",
         "//pkg/sql/types",
         "//pkg/sql/unsafesql",
         "//pkg/sql/vecindex",
